==> 02-data.Rmd <==
# Data sources


This is an overview of all data sources for this project. The multivariate analysis conducted in the rest of the book is solely based on these data sets.

### Population {.unlisted .unnumbered}

<u>Data description:</u> *Total population is based on the de facto definition of population, which counts all residents regardless of legal status or citizenship.*

[Link to data.](https://data.worldbank.org/indicator/SP.POP.TOTL)


### GDP {.unlisted .unnumbered}

<u>Data description:</u> *Measure of a country's economic output computed by adding up the monetary value (in current US$) of all finished goods and services made within a country for a given year.*

[Link to data.](https://data.worldbank.org/indicator/NY.GDP.MKTP.CD)


### HDI {.unlisted .unnumbered}

<u>Data description:</u> *Composite index measuring average achievement in three basic dimensions of human development-a long and healthy life, knowledge and a decent standard of living.*

[Link to data.](http://hdr.undp.org/en/indicators/137506)


### Life expectancy {.unlisted .unnumbered}

<u>Data description:</u> *Mean length of life of an actual birth cohort (all individuals born in a given year).*

[Link to data.](https://data.worldbank.org/indicator/SP.DYN.LE00.IN) 


### Years of schooling {.unlisted .unnumbered}

<u>Data description:</u> *Number of years of schooling that a child of school entrance age can expect to receive if prevailing patterns of age-specific enrollment rates persist throughout the child's life.*

[Link to data.](http://hdr.undp.org/en/indicators/69706#)


### Energy consumption {.unlisted .unnumbered}

<u>Data description:</u> *Energy consumption by source, country and year.*

[Link to data.](https://github.com/owid/energy-data)


### Country info {.unlisted .unnumbered}

<u>Data description:</u> *ISO code, region and level of income for all worldbank-recognized regions The .csv dataframe is in the same zip as the one containing GDP data (first source).*

[Link to data.](https://data.worldbank.org/indicator/NY.GDP.MKTP.CD)
