==> 03-cleaning.Rmd <==
# Data transformation

```{r, echo=FALSE, warning=FALSE}
library(dplyr)
library(tidyverse)
library(countrycode)
```

We aim at merging all data into one tidy master data frame within which **each observation will correspond to one country and one year**.

## Selecting the time range

We hope to keep a sufficiently wide range so as to **observe clear change in energy consumption behavior across time**, while at the same time reducing missing data for as many countries as possible.


```{r}
# Load all data sets
df_population_wide <- read.csv("data/population.csv", skip = 4, check.names = FALSE)
df_GDP_wide <- read.csv("data/worldbank_GDP.csv", skip = 4, check.names = FALSE)
df_HDI_wide <- read.csv("data/HDI.csv", skip = 5, check.names = FALSE)
df_life_wide <- read.csv("data/life_expectancy.csv", skip = 4, check.names = FALSE)
df_schooling_wide <- read.csv("data/un_years_of_schooling.csv", skip = 6, check.names = FALSE)
df_energy_long <- read.csv("data/owid_energy_data.csv", header = TRUE, stringsAsFactors = FALSE)
df_country_info <- read.csv("data/worldbank_country_info.csv", header = TRUE, stringsAsFactors = FALSE, encoding = "UTF-8")
```

```{r}
year_availability <- matrix(NaN, 6, 2)
colnames(year_availability) <- c("Min year", "Max year")
rownames(year_availability) <- c("Energy consumption" , "Years of schooling" , "Life expectancy" , "HDI" , "GDP", "Population")

year_availability[6,] <- c(1960, 2020)
year_availability[5,] <- c(1960, 2020)
year_availability[4,] <- c(1990, 2019)
year_availability[3,] <- c(1960, 2020)
year_availability[2,] <- c(1990, 2019)
year_availability[1,] <- c(1900, 2020)

## ncol and nrow
nr <- nrow(year_availability)
nc <- ncol(year_availability)

## separation between bars
sepr <- 0.2

op <- par(mar = c(5,9,4,2) + 0.1)

plot(c(1900,2020), c(1,nr), type="n",
     ylim=c(1-(1-sepr)/2-sepr,
            nr+(1-sepr)/2+sepr),
     ylab="", yaxt="n",
     xlab="", main = "Time range for each data set")
grid(ny = NA)

## plot bars
for(i in 1:nr) {
    lo <- year_availability[i,1]
    hi <- year_availability[i,2]
    polygon(x=c(lo, lo, hi, hi),
            y=c(i-(1-sepr)/2, i+(1-sepr)/2,
                i+(1-sepr)/2, i-(1-sepr)/2), col = "#6495ED")
}

## create y-axis
axis(2, at=1:nr, las=2, xpd=NA, labels=rownames(year_availability)) 
legend("topleft", fill="#6495ED", bty="n", legend="year is recorded")
## add center line
abline(v=0, lty=2)
```

We choose to restrict our multivariate analysis to the narrowest time range. This is exactly the intersection of all ranges : **1990-2019**.

We note that this plot **does not tell us anything about missing values**. Indeed, a given observation might be missing although within the time range of the data set.

* * *

## Selecting the variables

We hope to be able to analyze and understand the cross-correlation between energy consumption and human development. To do so, we need the following:

- Data indicators for **HDI and all its major components**, namely: `GDP`, `life_expectancy`, `years_of_schooling`;
- Figures for **consumption per energy type**;
- Population data and other country indicators like geographic region.

## Data transformation procedure

Our data transformation works as follows.

1. Determine the needed variables from the respective datasets 
2. Create dataframes for each of the datasets that only contains the needed variables and time frames to avoid having an unnecessarily large final dataframe that would slow down analyses
  - Energy data
  - GDP data
  - Life expectancy data
  - Schooling data
  - Population data
  - HDI data
3. Filter the dataframes on common ISO codes (unique identifiers for countries) such that only countries for which all dataframes match up are selected. We find that there is little to no variance in the countries from the different datasets but that some include geographic regions that are filtered by this procedure as they do not have ISO codes.
4. Convert previously wide dataframes into a long format such that each row is one unique observation of country and year
5. Merge all dataframes on their ISO code and year
6. Add the country names from the library `countrycode` based on the ISO codes to guarantee a common naming scheme and simplify interpretation.

We use the resulting `master_df` as our database in all further analyses. Of course, we sometimes have to modify or calculate new variables based on the current question.

```{r}
# `variable` includes all variables except for year and country
variable <- c("ISO_code", "HDI", "population", 
             "years_of_schooling", "life_expectancy", "GDP", "fossil_fuel", "hydro", "low_carb", 
               "nuclear", "other_renew", "renew", "solar", "wind", 
               "coal", "gas", "oil", "biofuel", "primary")
dataset <- c("worldbank_country_info", "HDI", "population", "un_years_of_schooling", "life_expectancy",
             "worldbank_gdp", "owid_energy_data", "owid_energy_data", "owid_energy_data", "owid_energy_data", "owid_energy_data", "owid_energy_data", "owid_energy_data", "owid_energy_data", "owid_energy_data", "owid_energy_data", "owid_energy_data", "owid_energy_data", "owid_energy_data")

chosen_variables <- data.frame(variable, dataset)

knitr::kable(chosen_variables, "html")
```

```{r, warning=FALSE}
# Choose years
relevant_years <- as.character(1990:2019)

df_energy_long <- df_energy_long %>% 
  select(-contains(c("_per_capita", "_pct", "_change", "_production", "_share"))) %>%
  subset(year %in% relevant_years) %>%
  rename("ISO_code" = "iso_code") %>%
  subset(ISO_code != "") %>%
  rename("fossil_fuel" = "fossil_fuel_consumption", 
         "hydro" = "hydro_consumption", 
         "low_carb" = "low_carbon_consumption", 
         "nuclear" = "nuclear_consumption", 
         "other_renew" = "other_renewable_consumption", 
         "renew" = "renewables_consumption", 
         "solar" = "solar_consumption", 
         "wind" = "wind_consumption", 
         "coal" = "coal_consumption", 
         "gas" = "gas_consumption", 
         "oil" = "oil_consumption", 
         "biofuel" = "biofuel_consumption", 
         "primary" = "primary_energy_consumption") %>%
  select("ISO_code", "year", "fossil_fuel", "hydro", "low_carb", "nuclear", "other_renew", "renew", 
         "solar", "wind", "coal", "gas", "oil", "biofuel", "primary")

df_energy_ISO_with_no_missing_year <- df_energy_long %>%
  select("ISO_code", "year") %>%
  group_by(ISO_code) %>%
  summarize(count = n()) %>%
  filter(count == 2019 - 1990 + 1) %>%
  pull(ISO_code)

df_energy_long <- df_energy_long %>%
  filter(ISO_code %in% df_energy_ISO_with_no_missing_year)



df_GDP_wide <- df_GDP_wide %>%
  select("Country Name", "Country Code", relevant_years) %>%
  rename("country" = "Country Name", "ISO_code" = "Country Code") %>%
  mutate_at(relevant_years, as.numeric) %>%
  select("ISO_code", relevant_years)
  


df_life_wide <- df_life_wide %>%
  select("Country Name", "Country Code", relevant_years) %>%
  rename("country" = "Country Name", "ISO_code" = "Country Code") %>%
  mutate_at(relevant_years, as.numeric) %>%
  select("ISO_code", relevant_years)


df_schooling_wide <- df_schooling_wide %>%
  select("Country", relevant_years) %>%
  mutate("ISO_code" = countrycode(Country, origin = "country.name", destination = "iso3c")) %>%
  mutate_at(relevant_years, as.numeric) %>%
  select("ISO_code", relevant_years)


df_population_wide <- df_population_wide %>%
  select("Country Name", "Country Code", relevant_years) %>%
  rename("country" = "Country Name", "ISO_code" = "Country Code") %>%
  mutate_at(relevant_years, as.numeric) %>%
  select("ISO_code", relevant_years)


df_HDI_wide <- df_HDI_wide %>%
  select("Country", relevant_years) %>%
  mutate("ISO_code" = countrycode(Country, origin = "country.name", destination = "iso3c")) %>%
  mutate_at(relevant_years, as.numeric) %>%
  select("ISO_code", relevant_years)


df_country_info <- as.data.frame(df_country_info) %>% 
  rename("ISO_code" = "X.U.FEFF.iso_code", "region" = "Region", "income_group" = "IncomeGroup", "special_notes" = "SpecialNotes")


# retrieve common countries across data sets based on ISO code
common_ISO_codes <- Reduce(intersect, list(df_energy_long$ISO_code,df_life_wide$ISO_code, df_schooling_wide$ISO_code, 
                                           df_GDP_wide$ISO_code, df_population_wide$ISO_code, df_HDI_wide$ISO_code))

# filter data sets on common ISO codes
df_energy_long <- df_energy_long %>% filter(ISO_code %in% common_ISO_codes)
df_GDP_wide <- df_GDP_wide %>% filter(ISO_code %in% common_ISO_codes)
df_life_wide <- df_life_wide %>% filter(ISO_code %in% common_ISO_codes)
df_schooling_wide <- df_schooling_wide %>% filter(ISO_code %in% common_ISO_codes)
df_population_wide <- df_population_wide %>% filter(ISO_code %in% common_ISO_codes)
df_HDI_wide <- df_HDI_wide %>% filter(ISO_code %in% common_ISO_codes)
```


```{r}
# Convert data frames to long format
df_GDP_long <- df_GDP_wide %>% 
  pivot_longer(cols = -c("ISO_code"), names_to = "year", values_to = "GDP")
df_schooling_long <- df_schooling_wide %>% 
  pivot_longer(cols = -c("ISO_code"), names_to = "year", values_to = "years_of_schooling") %>%
  mutate_at("year", as.numeric)
df_life_long <- df_life_wide %>% pivot_longer(cols = -c("ISO_code"), names_to = "year", values_to = "life_expectancy")
df_population_long <- df_population_wide %>% 
  pivot_longer(cols = -c("ISO_code"), names_to = "year", values_to = "population") 
df_HDI_long <- df_HDI_wide %>% pivot_longer(cols = -c("ISO_code"), names_to = "year", values_to = "HDI")
```

```{r}
# merge data frames into one master data frame
master_df <- merge(df_energy_long, merge(df_GDP_long, merge(df_life_long, merge(df_schooling_long, merge(df_population_long, df_HDI_long,
                by = c("ISO_code", "year")), by = c("ISO_code", "year")), by = c("ISO_code", "year")), 
                by = c("ISO_code", "year")), by = c("ISO_code", "year"))

# Add one variable for country name and reorder columns
master_df <- master_df %>%
  mutate(country = countrycode(ISO_code, origin = "iso3c", destination = "country.name")) %>%
  select(ISO_code, country, year, population, GDP, HDI, life_expectancy, years_of_schooling, everything())
```

```{r}
# write master_df to csv
write.csv(master_df,"data/master.csv", row.names = FALSE)
```
