==> 05-results.Rmd <==
# Results

```{r, echo=FALSE, warning=FALSE}
library(dplyr)
library(tidyverse)
library(patchwork)
library(Lock5withR)
library(ggpubr)
library(stringr)
library(esquisse)
library(gganimate)
library(ggrepel)
library(GGally)
library(parcoords)
library(tools)
library(grid)
library(leaflet)
library(rgdal)
library(RColorBrewer)
library(plotly)
```

```{r echo=FALSE, warning=FALSE}
# Load Data
master_df <- read.csv(file = "data/master.csv")
# Load country info data
df_country_info <- read.csv("data/worldbank_country_info.csv", header=TRUE, stringsAsFactors=FALSE, encoding = "UTF-8") %>%
    rename(c(ISO_code = "X.U.FEFF.iso_code", region="Region", income_group="IncomeGroup")) %>%
    select(c("ISO_code", "region", "income_group"))

master_df <- left_join(master_df, df_country_info, by = "ISO_code")
```

```{r}
energy_vars <- c("fossil_fuel", "hydro", "low_carb", "nuclear", "other_renew", "renew", "solar", "wind", "coal", "gas", "oil", "biofuel", "primary")
distinct_energy_vars <- c("hydro", "nuclear", "solar", "wind", "coal", "gas", "oil", "biofuel", "other_renew")
```

## Cross directional causality between energy consumption (EC), GDP and years of schooling (YOS)

The acronyms EC, GDP and YOS will be used to refer to energy consumption, gross domestic product and years of schooling respectively.

### Energy consumption fuels economic prosperity {.unlisted .unnumbered}

Analyzing the impact of energy consumption on GDP and mean years of schooling is not that easy as the three variables suffer from cross directional causality.
The intuition suggests a **high correlation between energy consumption and GDP**. Indeed, most economies are **highly energy dependent** and many production and consumption activities involve energy as a central input. 


```{r, fig.height=7, fig.width = 7}
# plot GDP per capita vs primary energy consumption per capita in 2019
master_df %>%
    filter(year == 2019) %>%
    ggplot(aes(x=primary/population, y=GDP/population)) +
    geom_point(aes(color = region, size = population, label = country)) +
    geom_smooth(method="lm", color = "black") +
    theme_light() +
    scale_x_log10(labels = function(l) {trans = l * 1000}) +
    scale_y_log10(labels = function(l) {trans = l / 1000}) +
    geom_text(
      aes(label=country), vjust = 2,
      size = 3,
      check_overlap = T
    ) +
    ggtitle("GDP per capita vs. primary energy consumption per capita in 2019") +
    xlab("Primary energy consumption per capita (MWh)") + 
    ylab("GDP per capita (k$)") +
    theme_classic() +
    theme(legend.position="bottom", legend.box = "vertical", 
          panel.grid.major.x = element_line( size=.1, color="gray" ),  
          panel.grid.major.y = element_line( size=.1, color="gray" )) +
    guides(colour = guide_legend(order = 1),
            size = guide_legend(order = 2),
           colour = guide_legend(override.aes = list(size=4)))+
    scale_color_brewer(palette="Set2")
```


This graph emphasizes the intuitive idea that **energy is required to stimulate growth**. More consumption generally results in an increase in the economic output.

* * *

### Parallel increase of EC and YOS in economic growth  {.unlisted .unnumbered}

As a country benefits from economic prosperity with a higher GDP per capita, this one might allocate more resources to education to foster the formation of human capital. This would **enhance the schooling system and lead to the increase of the mean year of schooling** for a given cohort.


```{r, fig.width = 7}
# plot GDP per capita vs primary energy consumption per capita in 2019
master_df %>%
    filter(country %in% c("Brazil", "Russia", "India", "China")) %>%
    ggplot(aes(x=primary/population, y=years_of_schooling, group = country, colour = country)) +
    geom_point(aes(label = year)) +
    geom_path(size = 1,
              arrow = arrow(type = "open", angle = 30, length = unit(0.15, "inches"))) +
    theme_light() +
    scale_x_log10(labels = function(l) {trans = l * 1000}) +
    ggtitle("Mean years of schooling vs. primary energy consumption per capita \nfor BRIC countries over 1990-2019") +
    xlab("Primary energy consumption per capita (MWh)") + 
    ylab("Expected years of schooling at birth (years)") +
    theme_classic() +
    theme(legend.position="bottom", legend.box = "vertical", 
          panel.grid.major.x = element_line( size=.1, color="gray" ),  
          panel.grid.major.y = element_line( size=.1, color="gray" )) +
    guides(colour = guide_legend(order = 1),
           colour = guide_legend(override.aes = list(size=4))) +
  scale_colour_discrete(limits = c("Brazil", "Russia", "India", "China"))
```

```{r, fig.width = 7}
# plot GDP per capita vs primary energy consumption per capita in 2019
master_df %>%
    filter(country %in% c("United States", "Canada", "Japan", "France", "Germany")) %>%
    ggplot(aes(x=primary/population, y=years_of_schooling, group = country, colour = country)) +
    geom_point(aes(label = year)) +
    geom_path(size = 1,
              arrow = arrow(type = "open", angle = 30, length = unit(0.15, "inches"))) +
    theme_light() +
    scale_x_log10(labels = function(l) {trans = l * 1000}) +
    ggtitle("Mean years of schooling vs. primary energy consumption per capita \nfor most developed countries over 1990-2019") +
    xlab("Primary energy consumption per capita (MWh)") + 
    ylab("Expected years of schooling at birth (years)") +
    theme_classic() +
    theme(legend.position="bottom", legend.box = "vertical", 
          panel.grid.major.x = element_line( size=.1, color="gray" ),  
          panel.grid.major.y = element_line( size=.1, color="gray" )) +
    guides(colour = guide_legend(order = 1),
           colour = guide_legend(override.aes = list(size=4))) +
  scale_colour_discrete(limits = c("United States", "Canada", "Japan", "France", "Germany")) 
  
```

```{r,  fig.width = 7, warning=FALSE}
# plot GDP per capita vs primary energy consumption per capita in 2019 for the two groups of countries
master_df %>%
    filter(country %in% c("United States", "Canada", "Japan", "France", "Germany", "Brazil", "Russia", "India", "China")) %>%
    mutate(category = if_else(country %in% c("Brazil", "Russia", "India", "China"), "BRIC", "mature")) %>%
    ggplot(aes(x=primary/population, y=years_of_schooling, group = country, colour = category)) +
    geom_point(aes(label = year)) +
    geom_hline(yintercept = 15.0, color = "black", size = 0.5) +
    geom_hline(yintercept = 17.0, color = "black", size = 0.5) +
    annotate("rect", fill = "grey", alpha = 0.4, 
        xmin = 0.000, xmax = Inf, ymin = 15, ymax = 17) +
    annotate("text", x = 0.000005, y = 16, label = "YOS at maturity") +
    geom_path(size = 0.2) +
    theme_light() +
    geom_point(aes(group = seq_along(year))) +
    transition_reveal(year) +
    scale_x_log10(labels = function(l) {trans = l * 1000}) +
    ggtitle("Mean years of schooling vs. primary energy consumption per capita \n for BRIC and most developed countries over 1990-2019") +
    xlab("Primary energy consumption per capita (MWh)") + 
    ylab("Expected years of schooling at birth (years)") +
    theme_classic() +
    theme(legend.position="bottom", legend.box = "vertical", legend.title = element_blank(),
          panel.grid.major.x = element_line( size=.1, color="gray" ),  
          panel.grid.major.y = element_line( size=.1, color="gray" )) +
    guides(colour = guide_legend(order = 1),
           colour = guide_legend(override.aes = list(size=4))) +
  scale_color_manual(labels = c("BRIC : Brazil, Russia, India and China", "Mature countries : United States, Canada, Japan, France, Germany"), values=c("black", "#E69F00"))
```

The path towards economic maturity consists of a **simultaneous rise of energy consumption and mean years of schooling**. To become more productive, a country invests in education in order for its proportion of educated workers to increase. On the plot just above, we notice that **the expect years of schooling seems to converge towards a value between 15 and 17 years, value that we observe for mature countries**.

* * *

## Co-movement of energy consumption and HDI

### An overview of HDI in the world {.unlisted .unnumbered}

```{r, echo=FALSE}
# What countries have a high share of renewables today?
# Use 2019 percentage data
graph_df <- master_df

#Read shape file
world_spdf <- readOGR( 
  dsn= paste0(getwd(), "/R/world_shape_file"), 
  layer="TM_WORLD_BORDERS_SIMPL-0.3",
  verbose=FALSE
)

world_spdf@data = data.frame(world_spdf, graph_df[match(world_spdf$ISO3, graph_df$ISO_code),])

# Create a color palette for the map:
mybins <- rev(c(0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1.0))
mypalette <- colorBin( palette="RdYlGn", domain = graph_df$renew, na.color="#d3d3d3", bins=mybins)

# Display map
leaflet(world_spdf) %>% 
  addTiles()  %>% 
  setView( lat=10, lng=0 , zoom=2) %>%
  addPolygons( 
    fillColor = ~mypalette(HDI), 
    stroke=TRUE, 
    fillOpacity = 0.9, 
    color="white", 
    weight=0.3,
    labelOptions = labelOptions( 
      style = list("font-weight" = "normal", padding = "3px 8px"), 
      textsize = "13px", 
      direction = "auto"
    )
  ) %>%
  addLegend(pal=mypalette, values=~HDI, opacity=0.9, title = "HDI", position = "bottomleft" )
```
</br>
From the available data, we observe **high variability of the human development index**. It ranges from the highest values in Europe, North America, Oceania and Japan to the lowest in central Africa.

* * *

### Segmented analysis of HDI {.unlisted .unnumbered}

The increase of energy consumption of a population generally tends to affect its way of living. The intuition suggests that, in the long-term, consuming more energy is correlated with an improvement of the the human development of a population and hence, an increase of the HDI.


```{r, fig.height=7, fig.width = 7}
# energy consumption per capita threshold for segmented regression
log_pcpc_sep <- 3.5e-5

regression_data_inf <- master_df %>%
    filter(year == 2019, primary/population <= log_pcpc_sep) %>%
    mutate(log_ratio_primary_population = log10(primary/population))

regression_data_sup <- master_df %>%
    filter(year == 2019, primary/population > log_pcpc_sep) %>%
    mutate(log_ratio_primary_population = log10(primary/population))

lm_inf <- lm(data = regression_data_inf, HDI ~ log_ratio_primary_population)
lm_sup <- lm(data = regression_data_sup, HDI ~ log_ratio_primary_population)

r_squared_inf <- summary(lm_inf)$r.squared
r_squared_sup <- summary(lm_sup)$r.squared

# plot HDI vs primary energy consumption per capita in 2019
master_df %>%
    filter(year == 2019) %>%
    ggplot(aes(x=primary/population, y=HDI)) +
    geom_point(aes(color = region, size = population, label = country)) +
    geom_vline(xintercept = log_pcpc_sep, linetype = "dashed") +
    geom_smooth(data = subset(master_df, 2e-6 < primary/population & primary/population < log_pcpc_sep), method="lm", color = "black") +
    geom_smooth(data = subset(master_df, primary/population > log_pcpc_sep), method="lm", color = "black") +
    theme_light() +
    scale_x_log10(labels = function(l) {trans = l * 1000}) +
    scale_y_log10() +
    geom_text(
      aes(label=country), vjust = 2,
      size = 3,
      check_overlap = T
    ) +
    annotate("text", x = 0.00002, y= 0.55, label=paste("R^2 == ", round(r_squared_inf, digits = 2)), parse = T) +
    annotate("text", x = 0.0002, y= 0.55, label=paste("R^2 == ", round(r_squared_sup, digits = 2)), parse = T)  +
    ggtitle("HDI vs. primary energy consumption per capita in 2019") +
    xlab("Primary energy consumption per capita (MWh)") + 
    ylab("HDI") +
    theme_classic() +
    theme(legend.position="bottom", legend.box = "vertical", 
          panel.grid.major.x = element_line( size=.1, color="gray" ),  
          panel.grid.major.y = element_line( size=.1, color="gray" )) +
    guides(colour = guide_legend(order = 1),
            size = guide_legend(order = 2),
           colour = guide_legend(override.aes = list(size=4))) +
    scale_color_brewer(palette="Set2")
```

There seems to be a **threshold of energy consumption per capita which separates two types of HDI behaviors**. For countries with low level of energy consumption, the **causal relationship between energy consumption and HDI is very strong**, as suggested by the high value of $R^2$. By contrast, there does not seem to be a correlation for countries with high level of energy consumption. Intuitively, the investments made by a country on the verge of becoming wealthier will generally have a **much larger effect on human development** (health system, education).

* * *

## Association of human development and renewable energy 
*(Disclaimer: There is no data on renewable energy available for 91/169 countries. These countries are mainly in Africa and/or low-income countries and thus limit the analysis for these areas.)*

The earlier analysis of GDP and energy consumption shows that economic growth, energy consumption, and human development are correlated until a threshold on the economic growth is reached. Considering the growing importance of renewable energy and the negative effects fossil energy sources can have on human livelihood, we must understand the **relation of renewable energy and human development**.
```{r}
graph_df <- master_df[master_df$year==2019 & master_df$country != "Iceland"  & master_df$country != "Norway",]

# Create graph for MWh of renewable energy per capita 
p1 <- ggplot(graph_df, aes(x=renew/population*1000000000, y=HDI)) +
  geom_point(aes(color = region)) +
  geom_smooth(method="lm") +
  theme_classic() + 
  theme(legend.position="none") +
  geom_text(
    label=graph_df$country, 
    nudge_x = 0, nudge_y = 0.02,
    size = 3,
    check_overlap = T
  ) +
  scale_x_log10() +
  labs(x="Renewable energy consumption\nper capita (MWh)", y="HDI", color="Region", caption = " ") + 
  scale_color_brewer(palette="Set2")

# Create graph for renewable energy share of total energy consumption
p2 <- ggplot(graph_df, aes(x=renew/primary*100, y=HDI)) +
  geom_point(aes(color = region)) +
  geom_smooth(method="lm") +
  theme_classic() + 
  theme(legend.position="none") +
  geom_text(
    label=graph_df$country, 
    nudge_x = 0, nudge_y = 0.02,
    size = 3,
    check_overlap = T
  ) +
  labs(x="Share of renewables to \n overall energy consumption (%)", y="", color="Region", caption = "Extreme outliers were removed.") +
  scale_color_brewer(palette="Set2")

# function to only grab the legend from a graph
get_only_legend <- function(plot) { 
  plot_table <- ggplot_gtable(ggplot_build(plot))  
  legend_plot <- which(sapply(plot_table$grobs, function(x) x$name) == "guide-box")
  legend <- plot_table$grobs[[legend_plot]] 
  return(legend)  
}

# Grab legend from first plot for final combined plot
p1_legend <- ggplot(graph_df, aes(x=renew/primary*100, y=HDI)) +
  geom_point(aes(color = region)) +
  theme_classic() +
  theme(legend.position = "bottom") +
  labs(color="Region") +
  scale_color_brewer(palette="Set2")

# Combine plots and legend
legend <- get_only_legend(p1_legend)
combined_plot <- gridExtra::arrangeGrob(p1, p2, ncol=2)
gridExtra::grid.arrange(combined_plot, legend, nrow=2, heights = c(9, 2), top = textGrob("HDI score vs. renewable energy consumption per capita \n and as share of overall energy consumption in 2019")) 
```
</br>
There is a **positive correlation between HDI and renewable energy consumption** both in per capita and in relative terms. However, we can also see that countries with a low share of renewable energy consumption like Hong Kong or Australia achieve high levels of human development. Generally, we can see that on a percentage-basis the most renewable countries are either in Europe or Latin America. Additionally, the left plot supports the idea that higher overall energy consumption correlates with human development. Nevertheless, we can deduct that more developed countries also invest more in renewable energy.

***

The following graph show the relationships, correlations and distribution between the HDI score, life expectancy, years of schooling, and the share of renewable energy compared to overall energy consumption. 

```{r, fig.height=4}
perct_df <- master_df
perct_df[energy_vars] <- (perct_df[energy_vars] / rowSums(perct_df[distinct_energy_vars])) * 100
perct_df_2019 <- perct_df[perct_df$year==2019,]

graph_df <- perct_df_2019 %>% rename("Life expectancy (years)"="life_expectancy", "Years of schooling (years)"="years_of_schooling", "Renewables share (%)"="renew")

# Function to add smooting lines to scatter matrix
scat_mat_lr <- function(data, mapping, ...){
  p <- ggplot(data = data, mapping = mapping) + 
    geom_point() + 
    geom_smooth(method=lm, fill="blue", color="blue", ...)
  p
}

ggpairs(graph_df, columns = c(6, 7, 8, 14), lower = list(continuous = scat_mat_lr)) +
  labs(title = "Human development metrics and renewable energy in 2019") + 
  theme_classic() +
  theme(panel.border = element_rect(colour = "black", fill=NA, size=1))
```
</br>
Countries with a high share of renewable energy are also those countries in which citizens have a higher life expectancy and receive more education. While life expectancy is left-skewed and years of schooling looks normally distributed, the share of renewables is strongly right-skewed. This indicates great room for improvement in investments into renewable energy.

* * *

## Characteristics of countries with a high share of renewable energy 

Now, we want to analyse countries that have a high share of renewable energy and what differentiates them from the rest of the world.

**Countries and the share of renewable energy in their energy mixes**
```{r}
# What countries have a high share of renewables today?
# Use 2019 percentage data
graph_df <- perct_df_2019

#Read shape file
world_spdf <- readOGR( 
  dsn= paste0(getwd(), "/R/world_shape_file"), 
  layer="TM_WORLD_BORDERS_SIMPL-0.3",
  verbose=FALSE
)

world_spdf@data = data.frame(world_spdf, graph_df[match(world_spdf$ISO3, graph_df$ISO_code),])

# Create a color palette for the map:
mybins <- c(100,80, 60, 40, 30, 20, 10, 5)
mypalette <- colorBin(palette="Greens", domain = graph_df$renew, na.color="#d3d3d3", bins=mybins)
legend_text <- c("80 - 100", "60 - 80", "40 - 60", "30 - 40", "10 - 20", "5 - 10", "0 - 5", "No data available")

# Prepare the text for tooltips
labels <- lapply(seq(nrow(world_spdf@data)), function(i){
  if(is.na(world_spdf@data$renew[i])){
    paste(
      "<b>", "Country: ", world_spdf@data$country[i],"</b>", 
      sep="")
  } else {
    paste(
      "<b>", "Country: ", world_spdf@data$country[i],"</b>","<br/>", 
      "Renewables: ", round(world_spdf@data$renew[i], 1), "%<br/>", 
      "Fossil Fuels: ", round(world_spdf@data$fossil_fuel[i], 1), "%<br/>", 
      "Nuclear: ", round(world_spdf@data$nuclear[i], 1), "%",
      sep="")
  }
})

# Display map
m <- leaflet(world_spdf) %>% 
  addTiles()  %>% 
  setView( lat=20, lng=0 , zoom=1.5) %>%
  addPolygons( 
    fillColor = ~mypalette(renew), 
    stroke=TRUE, 
    fillOpacity = 0.9, 
    color="#444444", 
    weight=0.3,
    label = lapply(labels, htmltools::HTML),
    labelOptions = labelOptions( 
      style = list("font-weight" = "normal", padding = "3px 8px"), 
      textsize = "13px", 
      direction = "auto"
    ),
    highlightOptions = highlightOptions(color = "white", weight = 2,
      bringToFront = TRUE)
  ) %>%
  addLegend(pal=mypalette, values=~(-renew), opacity=0.9, title = "Renewables (%)", position = "bottomleft", 
            labFormat = function(type, cuts, p){paste0(legend_text)})

m 
```
</br>
The countries with the highest share of renewable energy are Iceland, Norway, and Brazil. These and many other countries that have a high share of renewable energy have a long shoreline compared to the countries' sizes.

* * *
```{r}
# What is the energy mix of the most renewable countries?
graph_df <- master_df %>% filter(year==2019) %>%
  arrange(desc(renew/primary)) %>% slice(1:10) %>% 
  select(-c("primary", "fossil_fuel", "renew")) %>% 
  pivot_longer(distinct_energy_vars, names_to = "energy") %>%
  mutate(value_per_capita = value / population * 1000000) %>%
  mutate(energy = sub("other_renew", "Other Renewables", energy)) %>%
  mutate(energy = sub("hydro", "Hydropower", energy)) %>%
  mutate(energy = toTitleCase(energy))
levels <- names(sort(tapply(graph_df$value_per_capita, graph_df$energy, sum)))

ggplot(graph_df, aes(x=reorder(country, -value_per_capita), y=value_per_capita, fill=factor(energy, levels=levels))) +
  geom_bar(stat="identity") +
  theme_classic() + 
  theme(panel.grid.major.y = element_line()) + 
  theme(axis.text = element_text(angle = 45, vjust = 0, hjust = 1)) +
  scale_fill_manual(values=c("#eabc35", "#8d9e2b", "#7cabcb", "#645f55", "#c0b879", "#d0d890", "#b18781", "black", "#1e7991")) +
  labs(y="Energy mix per capita (MWh)", x="Country", fill="Energy source", title="Energy mixes of the 10 most renewable countries")
```
</br>
The most renewable countries rely on hydropower but also oil for a large share of their per capita energy consumption. While it is commonly assumed that renewable energies cannot reliably service a large energy demand, the example of Iceland indicates that this does not have to be true (of course considering political and geographical circumstances). Brazil and Ecuador are two poorer countries than the rest but manage to achieve a high share of renewable energy, further indicating that sustainable energy consumption is not only achievable for rich nations.    

* * *
```{r}
# Calculate average energy mix of top 10 most renewable countries and compare to the rest of the world

# Create dataframe for weighted average of energy of top 10 most renewable countries
top_ten_df <- perct_df_2019 %>% 
  arrange(desc(renew/primary)) %>% 
  slice(1:10) %>% 
  select(-c("primary", "fossil_fuel", "renew")) %>%
  pivot_longer(distinct_energy_vars, names_to = "energy") %>%
  group_by(energy) %>%
  summarise(value = sum(value)) %>%
  mutate(value = value/sum(value)) %>%
  mutate(name = "Top ten")

# Create dataframe for weighted average of energy rest of the world
average_df <- perct_df_2019 %>% 
  arrange(desc(renew/primary)) %>% 
  slice(11:nrow(.)) %>% 
  select(-c("primary", "fossil_fuel", "renew")) %>%
  pivot_longer(distinct_energy_vars, names_to = "energy") %>% drop_na(value) %>%
  group_by(energy) %>%
  summarise(value = mean(value)) %>%
  mutate(value = value/sum(value)) %>%
  mutate(name = "World average")

# Bind average_df and top_ten_df and rename values to have one graph to plot
graph_df <- rbind(top_ten_df, average_df) %>%
  mutate(energy = sub("other_renew", "Other Renewables", energy)) %>%
  mutate(energy = sub("hydro", "Hydropower", energy)) %>%
  mutate(energy = toTitleCase(energy))

# Plot Bar chart comparing the percentage energy of top ten ad rest of the world
ggplot(graph_df, aes(x=factor(name, levels=c("Top ten", "World average")), y=value, fill=factor(energy, levels=levels))) +
  geom_bar(stat="identity") +
  theme_classic() + 
  theme(panel.grid.major.y = element_line()) +
  geom_text(aes(label = paste0(round(value, 3)*100,"%")),
            position = position_stack(vjust = 0.5), size = 3, color="#ffffff", check_overlap = T) +
  scale_fill_manual(values=c("#eabc35", "#8d9e2b", "#7cabcb", "#645f55", "#c0b879","#d0d890", "#b18781", "black", "#1e7991")) +
  labs(y="Energy mix (%)", x="", fill="Energy source", title="Comparison of the energy mix of the most renewable countries\nand the rest of the world in 2019")
```
</br>
The **major difference** between the top ten countries and the rest of the world **is the share of hydropower**. It is 23% higher than the one of the rest of the world. Additionally, the **top ten countries use 24.6% less coal** in their energy mix. Interestingly, **wind, biofuel, and solar make up a very small share of the energy mix in all countries** despite them being the ones that many countries put the most political attention to. 

```{r}
# How does the share of renewable energy change in different countries?
graph_df <- master_df %>% filter(!is.na(renew))
graph_df$income_group[graph_df$country=="Venezuela"] <- "Lower middle income"
graph_df <- graph_df %>%
  group_by(income_group, year) %>% 
  summarise(renew = sum(renew), total = sum(primary)) %>%
  mutate(renew = renew / total*100) %>%
  mutate(income_group = factor(income_group, levels=c("High income", "Upper middle income", "Lower middle income")))
  
ggplot(graph_df, aes(x=year, y=renew, group=income_group, color=income_group)) +
  geom_line(key_glyph="timeseries", size=1) +
  theme_classic() +
  theme(panel.grid.major.y = element_line(), legend.position="bottom") +
  labs(x="Year", 
       y="Share of renewable energy (%)", 
       color="Income Groups", 
       caption="No data available for lowest income group.", 
       title="Weighted average growth of renewable energy by income group over 1990-2019") +
  scale_colour_discrete(labels = c('High income\n(>12,535 USD)', 
                                   'Upper middle income\n(4,046 - 12,535 USD)', 
                                   'Lower middle income\n(1,036- 3,995 USD)'))
```
</br>
Contrary, to common expectation, **not the richest countries but upper middle income countries grow their renewable energy quickest**. Instead, populous upper middle income countries like Brazil, Turkey, and China, record steeper growth. Nevertheless, many upper-middle income countries do not currently increase their share of renewable energy. Albeit the average growth is slightly slower for high income countries, **growth is more consistent across high income countries**. An upward trend for all high-income trend could be identified. Lower middle income countries have only seen a slight uptick in recent years. On the one hand, countries like India and Venezuela slowly drive up the share of renewables. On the other hand, **countries like Pakistan, Vietnam, or Sri Lanka have recently seen their renewable energy share drop drastically**.
