==> 06-interactive.Rmd <==
# Interactive component

```{r, echo=FALSE, warning=FALSE}
library(dplyr)
library(tidyverse)
```

```{r echo=FALSE, warning=FALSE}
# Load Data
master_df <- read.csv(file = "data/master.csv")
distinct_energy_vars <- c("hydro", "nuclear", "solar", "wind", "coal", "gas", "oil", "biofuel", "other_renew")

left_df <- master_df %>%
  select(c("ISO_code", "year", "HDI", "solar", "biofuel", "wind", "coal", "nuclear", "other_renew", "gas", "oil", "hydro")) %>%
  na.omit()

right_df <- left_df[left_df$year==2019,] %>% select(c("ISO_code", "HDI"))

interactive_df <- left_join(left_df, right_df, by="ISO_code") %>%
  mutate(HDI = HDI.y) %>%
  select(-c("HDI.x", "HDI.y")) %>%
  mutate(HDI = case_when(HDI <= 0.549 ~ 'Low',
                         (HDI <= 0.699 & HDI > 0.549) ~ 'Medium',
                         (HDI <= 0.799 & HDI > 0.699) ~ 'High',
                         (HDI <= 1 & HDI > 0.799) ~ 'Very high')) %>%
  select(-"ISO_code") %>%
  group_by(year, HDI) %>%
  summarise_all(list(sum)) %>%
  rename(Year="year", Solar="solar", Biofuel="biofuel", Wind="wind",
         Coal="coal", Nuclear="nuclear", 
         "Other_renewables"="other_renew", Gas="gas", Oil="oil",
          Hydropower="hydro")

write.csv(interactive_df,"data/interactive6.csv", row.names = FALSE)
```


## Interactive visualization on the energy mix by HDI level {.unlisted .unnumbered}

The interactive plot offers the following features :

- Selecting only a part of the y-axis to identify the smaller categories;
- Hovering over each element of the legend to highlight the variable in the graph;
- Selecting countries of which HDI level you want to examine.

<style>
  body {
    font-family: Verdana, Arial, sans-serif;
  }

  h1 {
    font-size: 1.2em;
    margin: 60px 0 5px 0;
  }

  svg {
    width: 1200px;
    margin-left: auto;
    margin-right: auto;
  }

  b {
    margin-left: 50px;
  }

  .wrapper {
    width: 800px;
    margin-left: 100px;
  }

  .HDI_category {
    width: 50px;
    text-align: center;
  }
</style>


<!-- Load d3.js -->
<script src="https://d3js.org/d3.v7.js"></script>

<!-- Create a div where the graph will take place -->



  <div id="my_dataviz"></div>
  <!-- Initialize a select button -->
  <b>Select the wanted HDI category: </b>
  <select id="HDI_category"></select>
</body>
<script>
  // set the dimensions and margins of the graph
  const margin = { top: 60, right: 230, bottom: 20, left: 50 },
    width = 800 - margin.left - margin.right,
    height = 500 - margin.top - margin.bottom;

  // append the svg object to the body of the page
  const svg = d3
    .select("#my_dataviz")
    .append("svg")
    .attr("width", width + margin.left + margin.right)
    .attr("height", height + margin.top + margin.bottom)
    .append("g")
    .attr("transform", `translate(${margin.left}, ${margin.top})`);

  // Parse the Data
  d3.csv(
    "https://raw.githubusercontent.com/AntoninVidon/energyhumandevelopment/main/data/interactive7.csv"
  ).then(function (data) {
    // List of groups (here I have one group per column)
    const allGroup = ["Medium", "High", "Very high"];

    // add the options to the button
    d3.select("#HDI_category")
      .selectAll("myOptions")
      .data(allGroup)
      .enter()
      .append("option")
      .text(function (d) {
        return d;
      }) // text showed in the menu
      .attr("value", function (d) {
        return d;
      }); // corresponding value returned by the button

    // store old data
    old_data = data
    data = old_data.filter(function (d) {
      return d.HDI == "High";
    });

    //////////
    // GENERAL //
    //////////

    // List of groups = header of the csv files
    //const keys = data.columns.slice(1)
    const keys = [
      "Coal",
      "Oil",
      "Gas",
      "Hydropower",
      "Nuclear",
      "Solar",
      "Biofuel",
      "Wind",
      "Other_renewables",
    ];

    // color palette
    const color = d3
      .scaleOrdinal()
      .domain(keys)
      .range([
        "#645f55",
        "black",
        "#b18781",
        "#1e7991",
        "#c0b879",
        "#eabc35",
        "#8d9e2b",
        "#7cabcb",
        "#d0d890",
      ]);

    //stack the data?
    const stackedData = d3.stack().keys(keys)(data);

    //////////
    // AXIS //
    //////////

    // Add X axis
    const x = d3
      .scaleLinear()
      .domain(
        d3.extent(data, function (d) {
          return d.Year;
        })
      )
      .range([0, width]);
    const xAxis = svg
      .append("g")
      .attr("transform", `translate(0, ${height})`)
      .call(d3.axisBottom(x).ticks(5).tickFormat(d3.format("d")));

    // Add X axis label:
    svg
      .append("text")
      .attr("text-anchor", "end")
      .attr("x", width)
      .attr("y", height + 40)
      .text("Time (year)");

    // Add Y axis label:
    svg
      .append("text")
      .attr("text-anchor", "end")
      .attr("x", 0)
      .attr("y", -20)
      .text("Energy consumption split for the selected HDI range (TWh)")
      .attr("text-anchor", "start");

    // Add Y axis
    const y = d3
      .scaleLinear()
      .domain([
        0,
        d3.max(data, function (d) {
          return (
            +d.Hydropower +
            +d.Nuclear +
            +d.Solar +
            +d.Wind +
            +d.Coal +
            +d.Gas +
            +d.Oil +
            +d.Biofuel +
            +d.Other_renewables
          );
        }) * 1.1,
      ])
      .range([height, 0]);
    const yAxis = svg
      .append("g")
      .attr("transform", `translate(0, 0)`)
      .call(d3.axisLeft(y).ticks(5));

    //////////
    // BRUSHING AND CHART //
    //////////

    // Add a clipPath: everything out of this area won't be drawn.
    const clip = svg
      .append("defs")
      .append("svg:clipPath")
      .attr("id", "clip")
      .append("svg:rect")
      .attr("width", width)
      .attr("height", height)
      .attr("x", 0)
      .attr("y", 0);

    // Add brushing
    const brush = d3
      .brushY() // Add the brush feature using the d3.brush function
      .extent([
        [0, 0],
        [width, height],
      ]) // initialise the brush area: start at 0,0 and finishes at width,height: it means I select the whole graph area
      .on("end", updateChart); // Each time the brush selection changes, trigger the 'updateChart' function

    // Create the scatter variable: where both the circles and the brush take place
    const areaChart = svg.append("g").attr("clip-path", "url(#clip)");

    // Area generator
    const area = d3.area()
      .x(function (d) {
        return x(d.data.Year);
      })
      .y0(function (d) {
        return y(d[0]);
      })
      .y1(function (d) {
        return y(d[1]);
      });

    // Show the areas
    areaChart
      .selectAll("mylayers")
      .data(stackedData)
      .join("path")
      .attr("class", function (d) {
        return "myArea " + d.key;
      })
      .style("fill", function (d) {
        return color(d.key);
      })
      .attr("d", area);

    // Add the brushing
    areaChart.append("g").attr("class", "brush").call(brush);

    let idleTimeout;
    function idled() {
      idleTimeout = null;
    }

    // A function that update the chart for given boundaries
    function updateChart(event, d) {
      extent = event.selection;

      // If no selection, back to initial coordinate. Otherwise, update X axis domain
      if (!extent) {
        if (!idleTimeout) return (idleTimeout = setTimeout(idled, 350)); // This allows to wait a little bit
        y.domain(
          d3.extent(data, function (d) {
            return d.Year;
          })
        );
      } else {
        y.domain([y.invert(extent[1]), y.invert(extent[0])]);
        areaChart.select(".brush").call(brush.move, null); // This remove the grey brush area as soon as the selection has been done
      }

      // Update axis and area position
      yAxis.transition().duration(1000).call(d3.axisLeft(y).ticks(5));
      areaChart.selectAll("path").transition().duration(1000).attr("d", area);
    }

    // A function that update the chart
    function update(selectedGroup) {
      //stack the data?

      // get new data
      data = old_data.filter(function (d) {
        return d.HDI == selectedGroup;
      });

      const stackedData = d3.stack().keys(keys)(data);

      // set new y domain according to the new data
      y.domain([
        0,
        d3.max(data, function (d) {
          return (
            +d.Hydropower +
            +d.Nuclear +
            +d.Solar +
            +d.Wind +
            +d.Coal +
            +d.Gas +
            +d.Oil +
            +d.Biofuel +
            +d.Other_renewables
          );
        }),
      ]);

      // update y axis
      yAxis.transition().duration(1000).call(d3.axisLeft(y).ticks(5));
      // update areaChart

      console.log(data);
      areaChart.selectAll("path").remove();

      // Show the areas
      areaChart
        .selectAll("mylayers")
        .data(stackedData)
        .join("path")
        .transition()
        .duration(1000)
        .attr("class", function (d) {
          return "myArea " + d.key;
        })
        .style("fill", function (d) {
          return color(d.key);
        })
        .attr("d", area);
    }

    //////////
    // HIGHLIGHT GROUP //
    //////////

    // What to do when one group is hovered
    const highlight = function (event, d) {
      // reduce opacity of all groups
      d3.selectAll(".myArea").style("opacity", 0.1);
      // expect the one that is hovered
      d3.select("." + d).style("opacity", 1);
    };

    // And when it is not hovered anymore
    const noHighlight = function (event, d) {
      d3.selectAll(".myArea").style("opacity", 1);
    };

    //////////
    // LEGEND //
    //////////

    // Add one dot in the legend for each name.
    const size = 20;

    svg
      .selectAll("myrect")
      .data([
      "Other_renewables",
      "Wind",
      "Biofuel",
      "Solar",
      "Nuclear",
      "Hydropower",
      "Gas",
      "Oil",
      "Coal"
    ])
      .join("rect")
      .attr("x", width + 10)
      .attr("y", function (d, i) {
        return 10 + i * (size + 5);
      }) // 100 is where the first dot appears. 25 is the distance between dots
      .attr("width", size)
      .attr("height", size)
      .style("fill", function (d) {
        return color(d);
      })
      .on("mouseover", highlight)
      .on("mouseleave", noHighlight);

    // Add one dot in the legend for each name.


    svg
      .selectAll("mylabels")
      .data([
      "Other_renewables",
      "Wind",
      "Biofuel",
      "Solar",
      "Nuclear",
      "Hydropower",
      "Gas",
      "Oil",
      "Coal"
    ])
      .join("text")
      .attr("x", width + 10 + size * 1.2)
      .attr("y", function (d, i) {
        return 10 + i * (size + 5) + size / 2;
      }) // 100 is where the first dot appears. 25 is the distance between dots
      .style("fill", function (d) {
        return color(d);
      })
      .text(function (d) {
        return d;
      })
      .attr("text-anchor", "left")
      .style("alignment-baseline", "middle")
      .on("mouseover", highlight)
      .on("mouseleave", noHighlight);

    // When the button is changed, run the updateChart function
    d3.select("#HDI_category").on("change", function (event, d) {
      // recover the option that has been chosen
      const selectedOption = d3.select(this).property("value");
      // run the updateChart function with this selected option
      update(selectedOption);
    });
  });
</script>

<br>

The stacked areas bring a fresh way of looking at evolution of the energy mix by HDI. We observe that **the rate at which energy consumption is evolving is significantly higher for countries with medium and high HDI**. Across all HDI categories, consumption of coal, oil and gas remains largely dominant. In the case of medium HDI countries, the staggering increase in energy consumption is almost solely supported by fossil energy sources. On the other hand, **the evolution of the consumption of non-renewable energy seem to slow down in high HDI countries**. 

The solar energy seem to only be mastered by the wealthiest countries; probably because this specific mean of production might require a lot of prior investments. Interestingly, the potential of this energy source might in fact be higher in countries in central Africa, which benefit from a high level of solar exposition.
