==> 07-conclusion.Rmd <==
# Conclusion

In conclusion, the study of human development and energy consumption is insightful and very relevant. Understanding how education, life expectancy, wealth, overall energy consumption, and the shift to renewable energies relate to each other today, is crucial to make decisions on how we shape future societies. These will consume more energy and should rely on more renewable energy without sacrificing quality of life.

## Answers

**How is the increase in energy consumption boosting the education system in order to cope with the industrial ”boom” generated by rapid economic growth ?**

State of the art has revealed that economic growth is fueled by energy consumption. Moreover, as a country moves forward on its path towards economic prosperity, analyses reveal that it tends to allocate more resources to enhancing its schooling system. This leads to its population being more qualified for complex activities that generate a much higher economic output. On the other hand, we saw that emerging countries are trying to catch-up in terms of education system as their mean year of schooling for a given birth cohort is converging towards a range of "mature" values, that we observe for economically stable countries.

**Is the rate at which the HDI increases with respect to energy consumption constant, or does it start to decrease after a specific GDP per capita ?**

We observe a threshold of energy consumption that separates two types of HDI behavior. Up to a certain level of consumption, the HDI continues to improve. However, if the population continues to consumes more and more energy, it will not succeed in converting it into a better human development. There is no clear relationship between both indicators after this threshold. Intuitively, the investments made by a country on the verge of becoming wealthier have a much larger effect on human development as they are more likely to impact the core components of human development such as health or education.


**Is an energy mix oriented towards renewable sources associated with good human development indicators ?**

Yes, countries that have a high share and consumption of renewable energy perform better in the categories of the HDI. Countries are richer and people get better education and live longer. while a strong causation in this relationship is unlikely as GDP is a strong confounding variable, we do see that those countries with high renewable energy exceed those countries that match their wealth but not their renewable energy sourcing.

**What characterizes the countries that have the most renewable energy mix ?**

Countries with a highly renewable energy mix get a high share of their energy from hydropower but also still require large amount of oil. Other countries should investigate the possibility of hydropower energy more closely when possible. Additionally, the case of Iceland suggests that even a very high energy consumption - which will concern a significant proportion of countries in the future - can be serviced by a highly renewable energy grid.
Both high income and upper middle income countries currently grow their share of renewable energy quickly, indicating that not only *the* richest countries can make their energy mix more sustainable. However, this does not neglect the fact that wealthier countries have more capacity to invest in renewables and that developing countries might need support in this area.

## Limitations

First, as with many analyses, one of our main constraints is the availability of data. There is no energy mix data for low income countries available and thus any analysis is limited to lower middle income and richer countries. This does not negate our findings but does not allow making conclusions on low income countries.
Second, to craft comprehensible analyses and visualizations, we needed to reduce the data we used either along the temporal or the geographic dimension. However, the raw data is complex and large as it includes both all countries and a time frame reaching back until 1900. This means that one might be able to find more complex relationships like long-term trends when examining both dimensions together. 

## Future directions

We are particularly interested in three future directions of this line of research. First, should institutions be able to gather more data on low income countries, how are their energy consumption and human development metrics developing and interacting? This is particularly interesting as we might be able to see patterns from developed countries again or discover new patterns. Second, which concrete policy recommendations can be given, when analyzing the drivers and influencing variables in countries that have achieved a highly renewable energy mix and high human development. Third, we are convinced that there lies great potential in predictive and inference-oriented approaches. With this, future researches could explore what the causality between the observed variable is and how trying to influence one might affect the other.

## Summary

In our report, we found out that while economic growth and better living standards are associated with higher energy consumption, this resource-driven growth stagnates at a certain level and that other measures need to be found for further growth. This fact encourages the idea that we should not only increase consumption and economic production further but also consider other measures to effectively improve education and life expectancy. Instead, nations can invest in greener technologies like Iceland, New Zealand, or Norway to grow beyond the limitations that the simple GDP-energy relation faces at some point.
